+++ castle/cms/static/patterns/sticky.js
@@ -17,26 +17,21 @@ define([
       window.addEventListener("load", function(event) {
         self.cookie_data = $.cookie("sticky-footer");
 
-        // Sets the cookie to indicate the footer was closed and won't re-appear.
-        $(".close").click(function() {
-          self.cookie_data = self.setCookie("closed-footer")
-        });
-
-        if (self.cookie_data !== "closed-footer") {
+        if (self.cookie_data !== "footer-shown") {
           $(".sticky-footer").show();
             // If cookie hasn't been set yet, do the rising animation.
             if (typeof(self.cookie_data) === "undefined") {
               $(".sticky-footer")
                 .css({ bottom: "-100px" })
                 .animate({ bottom: "0px" }, "slow");
-              self.cookie_data = self.setCookie("no-animation");
-          }
-      }
+            }
+          self.cookie_data = self.setCookie("footer-shown");
+        }
       });
     },
 
-    setCookie: function(args) {
-      return $.cookie("sticky-footer", args, { path: '/' });
+    setCookie: function(value) {
+      return $.cookie("sticky-footer", value, { path: '/' });
     },
   });
   return StickyFooter;